pyproject.toml
[project]
name = "kp3"
version = "0.1.0"
description = "KP3 - Knowledge Processing Pipeline with semantic search, world model extraction, and provenance tracking"
readme = "README.md"
license = "MIT"
requires-python = ">=3.12"
authors = [
    {name = "Mark Lubin", email = "mark@kairix.org"}
]
keywords = [
    "ai", "memory", "semantic-search", "embeddings", "pgvector",
    "knowledge-base", "rag", "llm", "agent", "postgresql"
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Database",
    "Typing :: Typed",
]
dependencies = [
    # Database
    "sqlalchemy[asyncio]>=2.0",
    "asyncpg>=0.29",
    "alembic>=1.13",
    "pgvector>=0.3",
    # Config & validation
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    # LLM & embeddings
    "anthropic>=0.40",
    "openai>=1.0",  # OpenAI embeddings + DeepSeek (OpenAI-compatible API)
    # CLI
    "click>=8.1",
    "python-dotenv>=1.2.1",
    "rich>=14.2.0",
    # HTTP client
    "httpx>=0.27",
    # Query service
    "fastapi>=0.115",
    "uvicorn>=0.34",
    "fastmcp>=2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "testcontainers[postgres]>=4.0",
    "httpx>=0.27",
    "ruff>=0.8",
    "pyright>=1.1",
]

[project.scripts]
kp3 = "kp3.cli:cli"
kp3-service = "kp3.query_service.main:main"
kp3-mcp = "kp3.query_service.mcp:main"

[project.urls]
Homepage = "https://github.com/marklubin/kp3"
Documentation = "https://github.com/marklubin/kp3#readme"
Repository = "https://github.com/marklubin/kp3.git"
Issues = "https://github.com/marklubin/kp3/issues"
Changelog = "https://github.com/marklubin/kp3/blob/main/CHANGELOG.md"

[build-system]
requires = ["uv_build"]
build-backend = "uv_build"

[tool.uv]
package = true

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # pyflakes
    "I",      # isort
    "ANN",    # flake8-annotations
    "ASYNC",  # flake8-async
    "S",      # flake8-bandit
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "T20",    # flake8-print
    "RUF",    # ruff-specific
]
ignore = [
    "S101",    # assert used (fine in tests)
    "B008",    # function call in default argument (required for FastAPI dependencies)
]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["ANN", "S101"]
"examples/*" = ["ANN", "T20"]  # Examples can use print and skip annotations
"src/kp3/services/search.py" = ["S608"]  # SQL built from internal strings, not user input

[tool.pyright]
pythonVersion = "3.12"
typeCheckingMode = "strict"
venvPath = "."
venv = ".venv"
reportMissingTypeStubs = false
exclude = ["tests/**", "alembic/**"]  # Tests can access private members, alembic has less strict types

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "session"
asyncio_default_test_loop_scope = "session"
testpaths = ["tests"]
markers = [
    "docker: marks tests as requiring Docker/Podman (deselect with '-m \"not docker\"')",
    "vllm: marks tests as requiring vLLM/GPU (skipped automatically if unavailable)",
]
